LUCENE-10600: SortedSetDocValues#docValueCount should be an int, not long (#960)

+++ lucene/core/src/java/org/apache/lucene/index/SingletonSortedSetDocValues.java
@@ -58,7 +58,7 @@ final class SingletonSortedSetDocValues extends SortedSetDocValues {
   }
 
   @Override
-  public long docValueCount() {
+  public int docValueCount() {
     return 1;
   }
